requirements.txt
asgiref==3.5.2
astroid==2.11.6
colorama==0.4.5
dill==0.3.5.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==2.4.0
pylint==2.14.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
tzdata==2022.1
wrapt==1.14.1
